all logic related to the node info bar has been moved to a seperate component
