+++ src/types.ts
@@ -2,6 +2,8 @@ import {Edge, Node} from "@xyflow/react";
 import Transition from "./classes/transition.ts";
 import StateMachine from "./classes/stateMachine.ts";
 import State from "./classes/state.ts";
+import React from "react";
+import StateOrStateMachineService from "./services/stateOrStateMachineService.tsx";
 
 // One Type to avoid repeating ourselves. Can be expanded/unionized as needed.
 export type CsmNodeProps = {state: State} | {stateMachine: StateMachine} | {name: string};
@@ -30,6 +32,21 @@ export type ExitNode = Node<{name: string}, 'exit-node'>;
 export type CsmEdgeProps = Edge<{
     transition: Transition
 }, 'csmEdge'>;
+
+
+export type ReactFlowContextProps = {
+    nodes: Node<CsmNodeProps>[]
+    setNodes: React.Dispatch<React.SetStateAction<Node<CsmNodeProps>[]>>;
+    edges: Edge[]
+    setEdges:  React.Dispatch<React.SetStateAction<Edge[]>>
+    selectedNode: Node<CsmNodeProps> | null
+    setSelectedNode: React.Dispatch<React.SetStateAction<Node<CsmNodeProps> | null>>
+    showSidebar: boolean
+    setShowSidebar:  React.Dispatch<React.SetStateAction<boolean>>
+    nameInput: string,
+    setNameInput: React.Dispatch<React.SetStateAction<string>>
+    stateOrStateMachineService: StateOrStateMachineService
+}
